# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Stipulate Card Benefit Intelligence API
  version: 1.0.0
  description: Parse stipulations. Route the payment.
  contact:
    name: Stipulate
    url: https://stipulate.io
servers:
  - url: https://api.stipulate.io/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
security:
  - ApiKeyAuth: []
paths:
...
  /webhooks:
    post:
      operationId: createWebhookSubscription
      summary: Subscribe to benefit change webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Webhook subscription created
...
  /keys:
    get:
      operationId: listApiKeys
      summary: List API keys for the authenticated org
      responses:
        '200':
          description: API key summaries (prefix only, no secrets)
    post:
      operationId: createApiKey
      summary: Create a new API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: New API key created (raw key returned once)
  /keys/{key_id}:
    delete:
      operationId: revokeApiKey
      summary: Revoke an API key
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Key revoked
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...